requirements.txt
nonebot2[fastapi,aiohttp,httpx,websockets]==2.3.3
pydantic>=2.7.0,<3.0.0
nonebot-adapter-console>=0.6.0,<1.0.0
nonebot-adapter-onebot>=2.4.4,<2.5.0
nonebot-adapter-qq>=1.5.1,<1.6.0
nonebot-adapter-telegram>=0.1.0b17,<0.2.0
nonebot-plugin-apscheduler>=0.5.0,<1.0.0
sqlalchemy>=2.0.30,<3.0.0
apscheduler>=3.10.4,<4.0.0
aiofiles>=24.0.0,<25.0.0
ujson>=5.10.0,<6.0.0
lxml>=5.3.0,<6.0.0
msgpack>=1.0.8,<2.0.0
numpy>=1.26.4,<2.0.0
matplotlib>=3.9.0,<4.0.0
pillow>=10.4.0,<11.0.0
imageio>=2.35.0,<3.0.0
psutil>=5.9.8,<6.0.0
pycryptodome>=3.20.0,<4.0.0
py7zr>=0.21.0,<1.0.0
pytz==2024.1
zhconv>=1.4.3,<2.0.0
rapidfuzz>=3.9.3,<4.0.0
emoji>=2.12.1,<3.0.0
openpyxl>=3.1.2,<4.0.0
qrcode[pil]==7.4.2
onedice==1.0.7